+++ src/gwt/src/org/rstudio/studio/client/application/DesktopFrame.java
@@ -180,6 +180,7 @@ public interface DesktopFrame extends JavaScriptPassthrough
    void setDisableRendererAccessibility(boolean disable);
    void getEnableSplashScreen(CommandWithArg<Boolean> callback);
    void setEnableSplashScreen(boolean enable);
+   void setShowWhatsNew(boolean enable);
 
    void setAutohideMenubar(boolean enable);